+++ lib/kamisaku/chrome_runner.rb
@@ -1,0 +1,18 @@
+module Kamisaku
+  class ChromeRunner
+    def initialize(sandbox: false, headless: true, gpu: false)
+      @base_args = ["--run-all-compositor-stages-before-draw"]
+      @base_args << "--no-sandbox" unless sandbox
+      @base_args << "--headless" if headless
+      @base_args << "--disable-gpu" unless gpu
+    end
+
+    def html_to_pdf(html_path, pdf_path, pdf_header_footer: false)
+      args = [*@base_args]
+      args << "--no-pdf-header-footer" unless pdf_header_footer
+      args << "--print-to-pdf=#{pdf_path}"
+      args << html_path
+      system("google-chrome", *args)
+    end
+  end
+end