+++ lib/kamisaku/errors.rb
@@ -1,0 +1,3 @@
+module Kamisaku
+  class Error < StandardError; end
+end